chore: split service into packages (WIP)
