--- backup/models.go
@@ -1,10 +1,0 @@
-package backup
-
-import (
-	"io"
-)
-
-type BackupService interface {
-	CreateBackup(unlockPassword string, w io.Writer) error
-	RestoreBackup(unlockPassword string, r io.Reader) error
-}
