+++ nip47/handle_balance_request.go
@@ -1,10 +1,9 @@
-package main
+package nip47
 
 import (
 	"context"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/sirupsen/logrus"
 )
@@ -13,7 +12,7 @@ const (
 	MSAT_PER_SAT = 1000
 )
 
-func (svc *Service) HandleGetBalanceEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandleGetBalanceEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
 	resp := svc.checkPermission(nip47Request, requestEvent.NostrId, app, 0)
 	if resp != nil {
@@ -32,22 +31,22 @@ func (svc *Service) HandleGetBalanceEvent(ctx context.Context, nip47Request *nip
 			"requestEventNostrId": requestEvent.NostrId,
 			"appId":               app.ID,
 		}).Infof("Failed to fetch balance: %v", err)
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
 		return
 	}
 
-	responsePayload := &nip47.BalanceResponse{
+	responsePayload := &BalanceResponse{
 		Balance: balance,
 	}
 
 	appPermission := db.AppPermission{}
-	svc.db.Where("app_id = ? AND request_method = ?", app.ID, nip47.PAY_INVOICE_METHOD).First(&appPermission)
+	svc.db.Where("app_id = ? AND request_method = ?", app.ID, PAY_INVOICE_METHOD).First(&appPermission)
 
 	maxAmount := appPermission.MaxAmount
 	if maxAmount > 0 {
@@ -55,7 +54,7 @@ func (svc *Service) HandleGetBalanceEvent(ctx context.Context, nip47Request *nip
 		responsePayload.BudgetRenewal = appPermission.BudgetRenewal
 	}
 
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
 		Result:     responsePayload,
 	}, nostr.Tags{})
