+++ nip47/handle_info_request.go
@@ -1,15 +1,14 @@
-package main
+package nip47
 
 import (
 	"context"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/sirupsen/logrus"
 )
 
-func (svc *Service) HandleGetInfoEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandleGetInfoEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
 	resp := svc.checkPermission(nip47Request, requestEvent.NostrId, app, 0)
 	if resp != nil {
@@ -29,10 +28,10 @@ func (svc *Service) HandleGetInfoEvent(ctx context.Context, nip47Request *nip47.
 			"appId":               app.ID,
 		}).Infof("Failed to fetch node info: %v", err)
 
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
@@ -45,7 +44,7 @@ func (svc *Service) HandleGetInfoEvent(ctx context.Context, nip47Request *nip47.
 		network = "mainnet"
 	}
 
-	responsePayload := &nip47.GetInfoResponse{
+	responsePayload := &GetInfoResponse{
 		Alias:       info.Alias,
 		Color:       info.Color,
 		Pubkey:      info.Pubkey,
@@ -55,7 +54,7 @@ func (svc *Service) HandleGetInfoEvent(ctx context.Context, nip47Request *nip47.
 		Methods:     svc.GetMethods(app),
 	}
 
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
 		Result:     responsePayload,
 	}, nostr.Tags{})
