+++ nip47/handle_list_transactions_request.go
@@ -1,17 +1,16 @@
-package main
+package nip47
 
 import (
 	"context"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/sirupsen/logrus"
 )
 
-func (svc *Service) HandleListTransactionsEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandleListTransactionsEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
-	listParams := &nip47.ListTransactionsParams{}
+	listParams := &ListTransactionsParams{}
 	resp := svc.decodeNip47Request(nip47Request, requestEvent, app, listParams)
 	if resp != nil {
 		publishResponse(resp, nostr.Tags{})
@@ -44,21 +43,21 @@ func (svc *Service) HandleListTransactionsEvent(ctx context.Context, nip47Reques
 			"appId":               app.ID,
 		}).Infof("Failed to fetch transactions: %v", err)
 
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
 		return
 	}
 
-	responsePayload := &nip47.ListTransactionsResponse{
+	responsePayload := &ListTransactionsResponse{
 		Transactions: transactions,
 	}
 
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
 		Result:     responsePayload,
 	}, nostr.Tags{})
