+++ nip47/handle_lookup_invoice_request.go
@@ -1,4 +1,4 @@
-package main
+package nip47
 
 import (
 	"context"
@@ -6,15 +6,14 @@ import (
 	"strings"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	decodepay "github.com/nbd-wtf/ln-decodepay"
 	"github.com/sirupsen/logrus"
 )
 
-func (svc *Service) HandleLookupInvoiceEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandleLookupInvoiceEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
-	lookupInvoiceParams := &nip47.LookupInvoiceParams{}
+	lookupInvoiceParams := &LookupInvoiceParams{}
 	resp := svc.decodeNip47Request(nip47Request, requestEvent, app, lookupInvoiceParams)
 	if resp != nil {
 		publishResponse(resp, nostr.Tags{})
@@ -45,10 +44,10 @@ func (svc *Service) HandleLookupInvoiceEvent(ctx context.Context, nip47Request *
 				"invoice":             lookupInvoiceParams.Invoice,
 			}).Errorf("Failed to decode bolt11 invoice: %v", err)
 
-			publishResponse(&nip47.Response{
+			publishResponse(&Response{
 				ResultType: nip47Request.Method,
-				Error: &nip47.Error{
-					Code:    nip47.ERROR_INTERNAL,
+				Error: &Error{
+					Code:    ERROR_INTERNAL,
 					Message: fmt.Sprintf("Failed to decode bolt11 invoice: %s", err.Error()),
 				},
 			}, nostr.Tags{})
@@ -66,21 +65,21 @@ func (svc *Service) HandleLookupInvoiceEvent(ctx context.Context, nip47Request *
 			"paymentHash":         lookupInvoiceParams.PaymentHash,
 		}).Infof("Failed to lookup invoice: %v", err)
 
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
 		return
 	}
 
-	responsePayload := &nip47.LookupInvoiceResponse{
+	responsePayload := &LookupInvoiceResponse{
 		Transaction: *transaction,
 	}
 
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
 		Result:     responsePayload,
 	}, nostr.Tags{})
