+++ nip47/handle_multi_pay_invoice_request.go
@@ -1,4 +1,4 @@
-package main
+package nip47
 
 import (
 	"context"
@@ -8,16 +8,15 @@ import (
 
 	"github.com/getAlby/nostr-wallet-connect/db"
 	"github.com/getAlby/nostr-wallet-connect/events"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	decodepay "github.com/nbd-wtf/ln-decodepay"
 	"github.com/sirupsen/logrus"
 )
 
 // TODO: pass a channel instead of publishResponse function
-func (svc *Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
-	multiPayParams := &nip47.MultiPayInvoiceParams{}
+	multiPayParams := &MultiPayInvoiceParams{}
 	resp := svc.decodeNip47Request(nip47Request, requestEvent, app, multiPayParams)
 	if resp != nil {
 		publishResponse(resp, nostr.Tags{})
@@ -29,7 +28,7 @@ func (svc *Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request
 	for _, invoiceInfo := range multiPayParams.Invoices {
 		wg.Add(1)
 		// TODO: we should call the handle_payment_request (most of this code is duplicated)
-		go func(invoiceInfo nip47.MultiPayInvoiceElement) {
+		go func(invoiceInfo MultiPayInvoiceElement) {
 			defer wg.Done()
 			bolt11 := invoiceInfo.Invoice
 			// Convert invoice to lowercase string
@@ -44,10 +43,10 @@ func (svc *Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request
 
 				// TODO: Decide what to do if id is empty
 				dTag := []string{"d", invoiceInfo.Id}
-				publishResponse(&nip47.Response{
+				publishResponse(&Response{
 					ResultType: nip47Request.Method,
-					Error: &nip47.Error{
-						Code:    nip47.ERROR_INTERNAL,
+					Error: &Error{
+						Code:    ERROR_INTERNAL,
 						Message: fmt.Sprintf("Failed to decode bolt11 invoice: %s", err.Error()),
 					},
 				}, nostr.Tags{dTag})
@@ -103,10 +102,10 @@ func (svc *Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request
 					},
 				})
 
-				publishResponse(&nip47.Response{
+				publishResponse(&Response{
 					ResultType: nip47Request.Method,
-					Error: &nip47.Error{
-						Code:    nip47.ERROR_INTERNAL,
+					Error: &Error{
+						Code:    ERROR_INTERNAL,
 						Message: err.Error(),
 					},
 				}, nostr.Tags{dTag})
@@ -125,9 +124,9 @@ func (svc *Service) HandleMultiPayInvoiceEvent(ctx context.Context, nip47Request
 					"amount": paymentRequest.MSatoshi / 1000,
 				},
 			})
-			publishResponse(&nip47.Response{
+			publishResponse(&Response{
 				ResultType: nip47Request.Method,
-				Result: nip47.PayResponse{
+				Result: PayResponse{
 					Preimage: response.Preimage,
 					FeesPaid: response.Fee,
 				},
