+++ nip47/handle_pay_keysend_request.go
@@ -1,18 +1,17 @@
-package main
+package nip47
 
 import (
 	"context"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
 	"github.com/getAlby/nostr-wallet-connect/events"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/sirupsen/logrus"
 )
 
-func (svc *Service) HandlePayKeysendEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
+func (svc *nip47Service) HandlePayKeysendEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
 
-	payParams := &nip47.KeysendParams{}
+	payParams := &KeysendParams{}
 	resp := svc.decodeNip47Request(nip47Request, requestEvent, app, payParams)
 	if resp != nil {
 		publishResponse(resp, nostr.Tags{})
@@ -28,10 +27,10 @@ func (svc *Service) HandlePayKeysendEvent(ctx context.Context, nip47Request *nip
 	payment := db.Payment{App: *app, RequestEvent: *requestEvent, Amount: uint(payParams.Amount / 1000)}
 	err := svc.db.Create(&payment).Error
 	if err != nil {
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
@@ -59,10 +58,10 @@ func (svc *Service) HandlePayKeysendEvent(ctx context.Context, nip47Request *nip
 				"amount":  payParams.Amount / 1000,
 			},
 		})
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
@@ -77,9 +76,9 @@ func (svc *Service) HandlePayKeysendEvent(ctx context.Context, nip47Request *nip
 			"amount":  payParams.Amount / 1000,
 		},
 	})
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
-		Result: nip47.PayResponse{
+		Result: PayResponse{
 			Preimage: preimage,
 		},
 	}, nostr.Tags{})
