+++ nip47/handle_sign_message_request.go
@@ -1,16 +1,15 @@
-package main
+package nip47
 
 import (
 	"context"
 
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/sirupsen/logrus"
 )
 
-func (svc *Service) HandleSignMessageEvent(ctx context.Context, nip47Request *nip47.Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*nip47.Response, nostr.Tags)) {
-	signParams := &nip47.SignMessageParams{}
+func (svc *nip47Service) HandleSignMessageEvent(ctx context.Context, nip47Request *Request, requestEvent *db.RequestEvent, app *db.App, publishResponse func(*Response, nostr.Tags)) {
+	signParams := &SignMessageParams{}
 	resp := svc.decodeNip47Request(nip47Request, requestEvent, app, signParams)
 	if resp != nil {
 		publishResponse(resp, nostr.Tags{})
@@ -34,22 +33,22 @@ func (svc *Service) HandleSignMessageEvent(ctx context.Context, nip47Request *ni
 			"requestEventNostrId": requestEvent.NostrId,
 			"appId":               app.ID,
 		}).Infof("Failed to sign message: %v", err)
-		publishResponse(&nip47.Response{
+		publishResponse(&Response{
 			ResultType: nip47Request.Method,
-			Error: &nip47.Error{
-				Code:    nip47.ERROR_INTERNAL,
+			Error: &Error{
+				Code:    ERROR_INTERNAL,
 				Message: err.Error(),
 			},
 		}, nostr.Tags{})
 		return
 	}
 
-	responsePayload := nip47.SignMessageResponse{
+	responsePayload := SignMessageResponse{
 		Message:   signParams.Message,
 		Signature: signature,
 	}
 
-	publishResponse(&nip47.Response{
+	publishResponse(&Response{
 		ResultType: nip47Request.Method,
 		Result:     responsePayload,
 	}, nostr.Tags{})
