+++ nip47/nip47_notifier.go
@@ -1,17 +1,18 @@
-// TODO: move to nip47
-package main
+package nip47
 
 import (
 	"context"
 	"encoding/json"
 	"errors"
 
+	"github.com/getAlby/nostr-wallet-connect/config"
 	"github.com/getAlby/nostr-wallet-connect/db"
 	"github.com/getAlby/nostr-wallet-connect/events"
-	"github.com/getAlby/nostr-wallet-connect/nip47"
+	"github.com/getAlby/nostr-wallet-connect/lnclient"
 	"github.com/nbd-wtf/go-nostr"
 	"github.com/nbd-wtf/go-nostr/nip04"
 	"github.com/sirupsen/logrus"
+	"gorm.io/gorm"
 )
 
 type Relay interface {
@@ -19,14 +20,22 @@ type Relay interface {
 }
 
 type Nip47Notifier struct {
-	svc   *Service
-	relay Relay
+	db       *gorm.DB
+	nip47Svc *nip47Service
+	relay    Relay
+	logger   *logrus.Logger
+	cfg      config.Config
+	lnClient lnclient.LNClient
 }
 
-func NewNip47Notifier(svc *Service, relay Relay) *Nip47Notifier {
+func NewNip47Notifier(nip47Svc *nip47Service, relay Relay, logger *logrus.Logger, cfg config.Config, db *gorm.DB, lnClient lnclient.LNClient) *Nip47Notifier {
 	return &Nip47Notifier{
-		svc:   svc,
-		relay: relay,
+		nip47Svc: nip47Svc,
+		relay:    relay,
+		logger:   logger,
+		cfg:      cfg,
+		db:       db,
+		lnClient: lnClient,
 	}
 }
 
@@ -35,40 +44,36 @@ func (notifier *Nip47Notifier) ConsumeEvent(ctx context.Context, event *events.E
 		return nil
 	}
 
-	if notifier.svc.lnClient == nil {
-		return nil
-	}
-
 	paymentReceivedEventProperties, ok := event.Properties.(*events.PaymentReceivedEventProperties)
 	if !ok {
-		notifier.svc.logger.WithField("event", event).Error("Failed to cast event")
+		notifier.logger.WithField("event", event).Error("Failed to cast event")
 		return errors.New("failed to cast event")
 	}
 
-	transaction, err := notifier.svc.lnClient.LookupInvoice(ctx, paymentReceivedEventProperties.PaymentHash)
+	transaction, err := notifier.lnClient.LookupInvoice(ctx, paymentReceivedEventProperties.PaymentHash)
 	if err != nil {
-		notifier.svc.logger.
+		notifier.logger.
 			WithField("paymentHash", paymentReceivedEventProperties.PaymentHash).
 			WithError(err).
 			Error("Failed to lookup invoice by payment hash")
 		return err
 	}
 
-	notifier.notifySubscribers(ctx, &nip47.Notification{
+	notifier.notifySubscribers(ctx, &Notification{
 		Notification:     transaction,
-		NotificationType: nip47.PAYMENT_RECEIVED_NOTIFICATION,
+		NotificationType: PAYMENT_RECEIVED_NOTIFICATION,
 	}, nostr.Tags{})
 	return nil
 }
 
-func (notifier *Nip47Notifier) notifySubscribers(ctx context.Context, notification *nip47.Notification, tags nostr.Tags) {
+func (notifier *Nip47Notifier) notifySubscribers(ctx context.Context, notification *Notification, tags nostr.Tags) {
 	apps := []db.App{}
 
 	// TODO: join apps and permissions
-	notifier.svc.db.Find(&apps)
+	notifier.db.Find(&apps)
 
 	for _, app := range apps {
-		hasPermission, _, _ := notifier.svc.hasPermission(&app, nip47.NOTIFICATIONS_PERMISSION, 0)
+		hasPermission, _, _ := notifier.nip47Svc.HasPermission(&app, NOTIFICATIONS_PERMISSION, 0)
 		if !hasPermission {
 			continue
 		}
@@ -76,15 +81,15 @@ func (notifier *Nip47Notifier) notifySubscribers(ctx context.Context, notificati
 	}
 }
 
-func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App, notification *nip47.Notification, tags nostr.Tags) {
-	notifier.svc.logger.WithFields(logrus.Fields{
+func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App, notification *Notification, tags nostr.Tags) {
+	notifier.logger.WithFields(logrus.Fields{
 		"notification": notification,
 		"appId":        app.ID,
 	}).Info("Notifying subscriber")
 
-	ss, err := nip04.ComputeSharedSecret(app.NostrPubkey, notifier.svc.cfg.GetNostrSecretKey())
+	ss, err := nip04.ComputeSharedSecret(app.NostrPubkey, notifier.cfg.GetNostrSecretKey())
 	if err != nil {
-		notifier.svc.logger.WithFields(logrus.Fields{
+		notifier.logger.WithFields(logrus.Fields{
 			"notification": notification,
 			"appId":        app.ID,
 		}).WithError(err).Error("Failed to compute shared secret")
@@ -93,7 +98,7 @@ func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App
 
 	payloadBytes, err := json.Marshal(notification)
 	if err != nil {
-		notifier.svc.logger.WithFields(logrus.Fields{
+		notifier.logger.WithFields(logrus.Fields{
 			"notification": notification,
 			"appId":        app.ID,
 		}).WithError(err).Error("Failed to stringify notification")
@@ -101,7 +106,7 @@ func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App
 	}
 	msg, err := nip04.Encrypt(string(payloadBytes), ss)
 	if err != nil {
-		notifier.svc.logger.WithFields(logrus.Fields{
+		notifier.logger.WithFields(logrus.Fields{
 			"notification": notification,
 			"appId":        app.ID,
 		}).WithError(err).Error("Failed to encrypt notification payload")
@@ -112,15 +117,15 @@ func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App
 	allTags = append(allTags, tags...)
 
 	event := &nostr.Event{
-		PubKey:    notifier.svc.cfg.GetNostrPublicKey(),
+		PubKey:    notifier.cfg.GetNostrPublicKey(),
 		CreatedAt: nostr.Now(),
-		Kind:      nip47.NOTIFICATION_KIND,
+		Kind:      NOTIFICATION_KIND,
 		Tags:      allTags,
 		Content:   msg,
 	}
-	err = event.Sign(notifier.svc.cfg.GetNostrSecretKey())
+	err = event.Sign(notifier.cfg.GetNostrSecretKey())
 	if err != nil {
-		notifier.svc.logger.WithFields(logrus.Fields{
+		notifier.logger.WithFields(logrus.Fields{
 			"notification": notification,
 			"appId":        app.ID,
 		}).WithError(err).Error("Failed to sign event")
@@ -129,13 +134,13 @@ func (notifier *Nip47Notifier) notifySubscriber(ctx context.Context, app *db.App
 
 	err = notifier.relay.Publish(ctx, *event)
 	if err != nil {
-		notifier.svc.logger.WithFields(logrus.Fields{
+		notifier.logger.WithFields(logrus.Fields{
 			"notification": notification,
 			"appId":        app.ID,
 		}).WithError(err).Error("Failed to publish notification")
 		return
 	}
-	notifier.svc.logger.WithFields(logrus.Fields{
+	notifier.logger.WithFields(logrus.Fields{
 		"notification": notification,
 		"appId":        app.ID,
 	}).Info("Published notification event")
