+++ service/models.go
@@ -3,8 +3,11 @@ package service
 import (
 	"github.com/getAlby/nostr-wallet-connect/alby"
 	"github.com/getAlby/nostr-wallet-connect/config"
-	"github.com/getAlby/nostr-wallet-connect/db"
+	"github.com/getAlby/nostr-wallet-connect/events"
 	"github.com/getAlby/nostr-wallet-connect/lnclient"
+	"github.com/getAlby/nostr-wallet-connect/nip47"
+	"github.com/sirupsen/logrus"
+	"gorm.io/gorm"
 )
 
 type Service interface {
@@ -14,7 +17,11 @@ type Service interface {
 	StopApp()
 	StopLNClient() error
 	StopDb() error
-	GetBudgetUsage(appPermission *db.AppPermission) int64
 	GetLogFilePath() string
 	GetAlbyOAuthSvc() alby.AlbyOAuthService
+	GetNip47Service() nip47.Nip47Service
+	GetLogger() *logrus.Logger
+	GetDB() *gorm.DB
+	WaitShutdown()
+	GetEventPublisher() events.EventPublisher
 }
