+++ tests/create_app.go
@@ -1,0 +1,28 @@
+package tests
+
+import (
+	"github.com/getAlby/nostr-wallet-connect/db"
+	"github.com/nbd-wtf/go-nostr"
+	"github.com/nbd-wtf/go-nostr/nip04"
+)
+
+func createApp(svc *TestService) (app *db.App, ss []byte, err error) {
+	senderPrivkey := nostr.GeneratePrivateKey()
+	senderPubkey, err := nostr.GetPublicKey(senderPrivkey)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	ss, err = nip04.ComputeSharedSecret(svc.cfg.GetNostrPublicKey(), senderPrivkey)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	app = &db.App{Name: "test", NostrPubkey: senderPubkey}
+	err = svc.db.Create(app).Error
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return app, ss, nil
+}
