+++ wails/wails_handlers.go
@@ -1,4 +1,4 @@
-package main
+package wails
 
 import (
 	"encoding/json"
@@ -12,7 +12,6 @@ import (
 	"github.com/getAlby/nostr-wallet-connect/alby"
 	"github.com/getAlby/nostr-wallet-connect/api"
 	"github.com/getAlby/nostr-wallet-connect/db"
-	"github.com/getAlby/nostr-wallet-connect/lsp"
 	"github.com/wailsapp/wails/v2/pkg/runtime"
 )
 
@@ -36,9 +35,9 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 	case len(authCodeMatch) > 1:
 		code := authCodeMatch[1]
 
-		err := app.svc.albyOAuthSvc.CallbackHandler(ctx, code)
+		err := app.svc.GetAlbyOAuthSvc().CallbackHandler(ctx, code)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -59,7 +58,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		pubkey := appMatch[1]
 
 		userApp := db.App{}
-		findResult := app.svc.db.Where("nostr_pubkey = ?", pubkey).First(&userApp)
+		findResult := app.svc.GetDB().Where("nostr_pubkey = ?", pubkey).First(&userApp)
 
 		if findResult.RowsAffected == 0 {
 			return WailsRequestRouterResponse{Body: nil, Error: "App does not exist"}
@@ -73,7 +72,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			updateAppRequest := &api.UpdateAppRequest{}
 			err := json.Unmarshal([]byte(body), updateAppRequest)
 			if err != nil {
-				app.svc.logger.WithFields(logrus.Fields{
+				app.svc.GetLogger().WithFields(logrus.Fields{
 					"route":  route,
 					"method": method,
 					"body":   body,
@@ -168,9 +167,9 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 	switch route {
 	case "/api/alby/me":
-		me, err := app.svc.albyOAuthSvc.GetMe(ctx)
+		me, err := app.svc.GetAlbyOAuthSvc().GetMe(ctx)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -179,9 +178,9 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		return WailsRequestRouterResponse{Body: me, Error: ""}
 	case "/api/alby/balance":
-		balance, err := app.svc.albyOAuthSvc.GetBalance(ctx)
+		balance, err := app.svc.GetAlbyOAuthSvc().GetBalance(ctx)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -195,14 +194,14 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		payRequest := &alby.AlbyPayRequest{}
 		err := json.Unmarshal([]byte(body), payRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
 			}).WithError(err).Error("Failed to decode request to wails router")
 			return WailsRequestRouterResponse{Body: nil, Error: err.Error()}
 		}
-		err = app.svc.albyOAuthSvc.SendPayment(ctx, payRequest.Invoice)
+		err = app.svc.GetAlbyOAuthSvc().SendPayment(ctx, payRequest.Invoice)
 		if err != nil {
 			return WailsRequestRouterResponse{Body: nil, Error: err.Error()}
 		}
@@ -219,7 +218,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			createAppRequest := &api.CreateAppRequest{}
 			err := json.Unmarshal([]byte(body), createAppRequest)
 			if err != nil {
-				app.svc.logger.WithFields(logrus.Fields{
+				app.svc.GetLogger().WithFields(logrus.Fields{
 					"route":  route,
 					"method": method,
 					"body":   body,
@@ -236,7 +235,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		resetRouterRequest := &api.ResetRouterRequest{}
 		err := json.Unmarshal([]byte(body), resetRouterRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -270,7 +269,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			openChannelRequest := &api.OpenChannelRequest{}
 			err := json.Unmarshal([]byte(body), openChannelRequest)
 			if err != nil {
-				app.svc.logger.WithFields(logrus.Fields{
+				app.svc.GetLogger().WithFields(logrus.Fields{
 					"route":  route,
 					"method": method,
 					"body":   body,
@@ -317,7 +316,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		redeemOnchainFundsRequest := &api.RedeemOnchainFundsRequest{}
 		err := json.Unmarshal([]byte(body), redeemOnchainFundsRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -340,13 +339,13 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		return WailsRequestRouterResponse{Body: *signMessageResponse, Error: ""}
 	// TODO: review naming
 	case "/api/instant-channel-invoices":
-		newInstantChannelRequest := &lsp.NewInstantChannelInvoiceRequest{}
+		newInstantChannelRequest := &api.NewInstantChannelInvoiceRequest{}
 		err := json.Unmarshal([]byte(body), newInstantChannelRequest)
-		newInstantChannelResponseResponse, err := app.api.GetLSPService().NewInstantChannelInvoice(ctx, newInstantChannelRequest)
+		newInstantChannelResponse, err := app.api.NewInstantChannelInvoice(ctx, newInstantChannelRequest)
 		if err != nil {
 			return WailsRequestRouterResponse{Body: nil, Error: err.Error()}
 		}
-		return WailsRequestRouterResponse{Body: *newInstantChannelResponseResponse, Error: ""}
+		return WailsRequestRouterResponse{Body: *newInstantChannelResponse, Error: ""}
 	case "/api/peers":
 		switch method {
 		case "GET":
@@ -359,7 +358,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			connectPeerRequest := &api.ConnectPeerRequest{}
 			err := json.Unmarshal([]byte(body), connectPeerRequest)
 			if err != nil {
-				app.svc.logger.WithFields(logrus.Fields{
+				app.svc.GetLogger().WithFields(logrus.Fields{
 					"route":  route,
 					"method": method,
 					"body":   body,
@@ -393,7 +392,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		res := WailsRequestRouterResponse{Body: *infoResponse, Error: ""}
 		return res
 	case "/api/alby/link-account":
-		err := app.svc.albyOAuthSvc.LinkAccount(ctx)
+		err := app.svc.GetAlbyOAuthSvc().LinkAccount(ctx)
 		if err != nil {
 			return WailsRequestRouterResponse{Body: nil, Error: err.Error()}
 		}
@@ -407,7 +406,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		backupReminderRequest := &api.BackupReminderRequest{}
 		err := json.Unmarshal([]byte(body), backupReminderRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -417,7 +416,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		err = app.api.SetNextBackupReminder(backupReminderRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -429,7 +428,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		changeUnlockPasswordRequest := &api.ChangeUnlockPasswordRequest{}
 		err := json.Unmarshal([]byte(body), changeUnlockPasswordRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -439,7 +438,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		err = app.api.ChangeUnlockPassword(changeUnlockPasswordRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -451,7 +450,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		startRequest := &api.StartRequest{}
 		err := json.Unmarshal([]byte(body), startRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -460,7 +459,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		err = app.api.Start(startRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -474,7 +473,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		setupRequest := &api.SetupRequest{}
 		err := json.Unmarshal([]byte(body), setupRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -483,7 +482,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		err = app.api.Setup(ctx, setupRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -495,7 +494,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		sendPaymentProbesRequest := &api.SendPaymentProbesRequest{}
 		err := json.Unmarshal([]byte(body), sendPaymentProbesRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -504,7 +503,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		sendPaymentProbesResponse, err := app.api.SendPaymentProbes(ctx, sendPaymentProbesRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -516,7 +515,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		sendSpontaneousPaymentProbesRequest := &api.SendSpontaneousPaymentProbesRequest{}
 		err := json.Unmarshal([]byte(body), sendSpontaneousPaymentProbesRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -525,7 +524,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		sendSpontaneousPaymentProbesResponse, err := app.api.SendSpontaneousPaymentProbes(ctx, sendSpontaneousPaymentProbesRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -537,7 +536,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		backupRequest := &api.BasicBackupRequest{}
 		err := json.Unmarshal([]byte(body), backupRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -550,7 +549,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			DefaultFilename: "nwc.bkp",
 		})
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -560,7 +559,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		backupFile, err := os.Create(saveFilePath)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -570,10 +569,10 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		defer backupFile.Close()
 
-		err = app.api.GetBackupService().CreateBackup(backupRequest.UnlockPassword, backupFile)
+		err = app.api.CreateBackup(backupRequest.UnlockPassword, backupFile)
 
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -585,7 +584,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		restoreRequest := &api.BasicRestoreWailsRequest{}
 		err := json.Unmarshal([]byte(body), restoreRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -598,7 +597,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 			DefaultFilename: "nwc.bkp",
 		})
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -608,7 +607,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		backupFile, err := os.Open(backupFilePath)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -618,9 +617,9 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 
 		defer backupFile.Close()
 
-		err = app.api.GetBackupService().RestoreBackup(restoreRequest.UnlockPassword, backupFile)
+		err = app.api.RestoreBackup(restoreRequest.UnlockPassword, backupFile)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -638,7 +637,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		getLogOutputRequest := &api.GetLogOutputRequest{}
 		err := json.Unmarshal([]byte(body), getLogOutputRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -647,7 +646,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		}
 		logOutputResponse, err := app.api.GetLogOutput(ctx, logType, getLogOutputRequest)
 		if err != nil {
-			app.svc.logger.WithFields(logrus.Fields{
+			app.svc.GetLogger().WithFields(logrus.Fields{
 				"route":  route,
 				"method": method,
 				"body":   body,
@@ -657,7 +656,7 @@ func (app *WailsApp) WailsRequestRouter(route string, method string, body string
 		return WailsRequestRouterResponse{Body: logOutputResponse, Error: ""}
 	}
 
-	app.svc.logger.WithFields(logrus.Fields{
+	app.svc.GetLogger().WithFields(logrus.Fields{
 		"route":  route,
 		"method": method,
 	}).Error("Unhandled route")
